+++ src/NUglify.Tests/JavaScript/Bugs.cs
@@ -229,5 +229,12 @@ namespace NUglify.Tests.JavaScript
         {
 	        TestHelper.Instance.RunTest("-rename:all");
         }
+        
+
+        [Test]
+        public void Bug214()
+        {
+	        TestHelper.Instance.RunTest("-rename:all");
+        }
     }
 }