+++ src/NUglify/JavaScript/Visitors/AnalyzeNodeVisitor.cs
@@ -2628,6 +2628,11 @@ namespace NUglify.JavaScript.Visitors
 
                                     if (removeIfUnreferenced)
                                     {
+                                        // dont remove `value` from setters even if unused, this is invalid
+	                                    if ((paramDecl.Binding is BindingIdentifier bi) && bi.Name == "value" &&
+	                                        node.FunctionType == FunctionType.Setter)
+		                                    continue;
+
                                         node.ParameterDeclarations.RemoveAt(ndx);
                                     }
                                 }