+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNext (unpublished)
+- Fixed bug with unused setter `value` parameters being removed
 
 ## v1.11.5 (10 December 2020)
 - Fixes bug with `<br>`s when HtmlToText()ing